Reorganization of the build configuration
The ax_pva_py.m4 script accepts --with-top=<path>, allowing the
configure script to be run from any directory.
The RELEASE.local and CONFIG_SITE.local files it generates are
simpler and the RELEASE entries are in the right order.

The top-level Makefile has a number of changes:
* Introduce a variable for the tools/autoconf path
* The 'make configure' target removes any existing RELEASE.local
and CONFIG_SITE.local files instead of making the user do that.
* 'make distclean' removes all built files except in DOC_DIR.
* 'make' builds in configure before src, CHECK_RELEASE enabled.
* Use $(RM) and $(RMDIR) for portability

Comments in the configure/RELEASE and configure/CONFIG_SITE files
describe what variables to set for manual configuration.

The src/pvaccess/Makefile has various cleanups to match the above
changes.