Add monitor thread synchronization.
In case monitorStart() and monitorStop() are being called close
together and before the monitor thread was able to connect, several
different segfaults were observed either in monitorStartThread()
function pointing to destroyed channel pointer or in stopMonitor()
using non-initialized pvaClientMonitorRequesterPtr.
The synchronization waits indefinetely for monitor thread to stop
before proceeding to stop the monitor. Timeout is already enforced
in the thread calling ::connect().